Refactored nodemailer code to be DRY
