+++ backend/routes/autoEmails.js
@@ -3,68 +3,17 @@
  * and emails in the database. All email formats can be seen seperated by functionality
  * inside the emails folder inside of the backend directory. 
  *
- * @summary   Code for routes involved with automated emails. 
+ * @summary   Code for routes involved with sending automated emails. 
  * @author    Amrit Kaur Singh, Dhanush Nanjunda Reddy
  */
 const express = require("express");
-const nodemailer = require("nodemailer");
-const Email = require("email-templates");
-const config = require("../config");
 const { body } = require("express-validator");
-const { findAllEmails } = require("../db/services/email");
+const { findAllEmails, findPrimaryEmail } = require("../db/services/email");
+const {sendEmail} = require("../routes/services/mailer");
 const { addOrder } = require("../db/services/order");
 const { isValidated } = require("../middleware/validation");
 
 const router = express.Router();
-
-// transporter object for nodemailer
-const transporter =
-  config.uweast.user === ""
-    ? null
-    : nodemailer.createTransport({
-        host: "smtp.gmail.com",
-        port: 465,
-        secure: true,
-        auth: config.uweast,
-      });
-
-// template based sender object
-const mail =
-  config.uweast.user === ""
-    ? null
-    : new Email({
-        transport: transporter,
-        send: true,
-        preview: false,
-      });
-
-/**
- * Populates given email template with locals and sends it to to_email.
- * All emails are sent from the email account specified in dotenv.
- *
- * @param {string} template - template email
- * @param {string} to_email - the email address being sent to
- * @param {string} locals - the email address being sent to
- * @param {object} res - the response
- * @returns {object} - mail object / err
- */
-async function sendEmail(template, to_email, locals, res) {
-  // sends email only if mail has been successfully setup
-  if (mail != null) {
-    await mail.send({
-      template: template,
-      message: {
-        from: config.uweast.user,
-        to: to_email,
-      },
-      locals: locals,
-    });
-    console.log(`Email ${template} has been sent to ${to_email}.`);
-  } else {
-    console.log(`Error: Email ${template} could not be sent to ${to_email}.`);
-    return res.status(500).send("Server err");
-  }
-}
 
 /**
  * Adds the order to the Orders DB and also sends order receipts to customer's specified email
@@ -170,7 +119,7 @@ router.post(
   async (req, res, next) => {
     try {
       // get all emails inside of Emails DB
-      const emails = await findAllEmails();
+      const emails = await findPrimaryEmail();
       // error if no emails exist
       if (!emails.length) {
         return res.status(500).json({ errors: [{ msg: "no emails found" }] });
